raftstore: extract raftlog_gc
