+++ components/raftstore/src/store/worker/mod.rs
@@ -1,3 +1,4 @@
 pub mod cleanup_sst;
 pub mod consistency_check;
 pub mod metrics;
+pub mod raftlog_gc;
